move readme previews to preview directory
I was previously using gh-pages branch to host images that I used for
inline previews in the readme. Github likes to change the raw url
structure and gh-pages url structures, which caused my inline preview
images to break.

Github now supports relative links for images within
the repository. The previews folder was created in the master branch to
house the preview images, which will referenced in the readme.md using
relative paths. huzzah