Fix rbenv init when only .bash_profile is sourced
Likely do to a seperate underlying issue, but for now this fixes anytime
.bash_profile is sourced and .bash_profile.d/*.bash files are not
sourced. Underlying issue is on my todo list to fix and should be a
simple redoing of the sourcing of bash_profile.d in bash_profile to use
glob instead of a for loop that uses ls and a wildcard (*.bash) to
source said files.