Add vim-polyglot plugin and load it last :squirrel:
When I first added this plugin to Vundle it started causing my my
colorscheme to not load initially with vim. After realizing I could
switch to it still after it loaded I concluded that the new plugin
addition, being the larger size plugin, was causing the colorscheme not
to load before it was invoked in the .vimrc. I moved the colorscheme
plugin to the top of the bundles.vim and moved vim-polyglot to the
bottom. :